+++ web/views/admin_view.ex
@@ -1,0 +1,4 @@
+defmodule Collaboration.AdminView do
+  use Collaboration.Web, :view
+
+end