+++ CHANGELOG.md
@@ -8,3 +8,9 @@
   *  Integrated Bootstrap v4, JQUERY and SASS (and Popper dependency)
   *  fixed flash message system
   *  finished register form including server-side validation
+
+#### 0.0.3 - 08/18/2017
+  *  Finished Login and Registration
+  *  Added Admin Interface
+  *  Added Settings (for change password is required)
+  *  In Admin Interface you can now change users data (no password required)